python_backend/requirements.txt
# Core Dependencies
Flask==3.0.0
Flask-CORS==4.0.0
python-dotenv==1.0.0

# LLM Providers
anthropic>=0.7.0
openai>=1.0.0
google-genai>=1.7.0

# Data & Config
pydantic>=2.0.0
pydantic-settings>=2.0.0

# Database
SQLAlchemy>=2.0.0
alembic>=1.13.0

# HTTP Client
httpx>=0.25.0

# MCP
mcp>=1.5.0

# Development Dependencies (optional)
# pytest>=7.4.0
# pytest-cov>=4.1.0
# pytest-asyncio>=0.21.0
# black>=23.0.0
# flake8>=6.0.0
# mypy>=1.5.0
